+++ apps/client/src/app/pages/markets/markets-page-routing.module.ts
@@ -1,0 +1,15 @@
+import { NgModule } from '@angular/core';
+import { RouterModule, Routes } from '@angular/router';
+import { AuthGuard } from '@ghostfolio/client/core/auth.guard';
+
+import { MarketsPageComponent } from './markets-page.component';
+
+const routes: Routes = [
+  { path: '', component: MarketsPageComponent, canActivate: [AuthGuard] }
+];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class MarketsPageRoutingModule {}