+++ apps/client/src/app/pages/markets/markets-page.module.ts
@@ -1,0 +1,13 @@
+import { CommonModule } from '@angular/common';
+import { CUSTOM_ELEMENTS_SCHEMA, NgModule } from '@angular/core';
+import { GfHomeMarketModule } from '@ghostfolio/client/components/home-market/home-market.module';
+
+import { MarketsPageRoutingModule } from './markets-page-routing.module';
+import { MarketsPageComponent } from './markets-page.component';
+
+@NgModule({
+  declarations: [MarketsPageComponent],
+  imports: [CommonModule, GfHomeMarketModule, MarketsPageRoutingModule],
+  schemas: [CUSTOM_ELEMENTS_SCHEMA]
+})
+export class MarketsPageModule {}