+++ libs/ui/src/lib/benchmark/benchmark.component.ts
@@ -1,4 +1,10 @@
-import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
+import {
+  ChangeDetectionStrategy,
+  Component,
+  Input,
+  OnChanges
+} from '@angular/core';
+import { locale } from '@ghostfolio/common/config';
 import { resolveMarketCondition } from '@ghostfolio/common/helper';
 import { Benchmark } from '@ghostfolio/common/interfaces';
 
@@ -8,11 +14,17 @@ import { Benchmark } from '@ghostfolio/common/interfaces';
   templateUrl: './benchmark.component.html',
   styleUrls: ['./benchmark.component.scss']
 })
-export class BenchmarkComponent {
-  @Input() benchmark: Benchmark;
+export class BenchmarkComponent implements OnChanges {
+  @Input() benchmarks: Benchmark[];
   @Input() locale: string;
 
   public resolveMarketCondition = resolveMarketCondition;
 
   public constructor() {}
+
+  public ngOnChanges() {
+    if (!this.locale) {
+      this.locale = locale;
+    }
+  }
 }